[back] feat : Member 엔티티 컬럼 추가 및 수정으로 관련된 기능도 수정

+++ back/src/main/java/com/example/cpsplatform/member/repository/MemberRepository.java
@@ -1,12 +1,11 @@
 package com.example.cpsplatform.member.repository;
 
-import com.example.cpsplatform.member.Member;
+import com.example.cpsplatform.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member,Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
 
     Optional<Member> findMemberByLoginId(String loginId);
